修复了创建信息时提示删除原来上传文件时失败的bug。
之前条件判断错误了，新创建的信息怎么原来会上传文件呢-_-
此外，把全局异常处理改换成了用@responseBody 返回json数据

+++ lab-webapp/src/main/java/cn/newgxu/lab/info/controller/InfoController.java
@@ -305,7 +305,10 @@ public class InfoController {
 
 	private void fileDelete(Information info) throws RuntimeException {
 		Information origin = infoService.find(info.getId());
-		Assert.notNull("对不起，您所请求的资源不存在！", origin);
+//		如果没有，就代表是新建文档，直接返回吧- -
+		if (origin == null) {
+			return;
+		}
 		if (origin.getDocUrl() != null) {
 			File f = new File(Config.UPLOAD_ABSOLUTE_DIR + origin.getDocUrl());
 			if (!f.delete()) {
